With the recent MFC of i4b 1.00.1, spppcontrol is no longer usable to configure isp<N> interfaces; ispppcontrol must be used instead.
To avoid misunderstandings, remove isp0 interfaces from the examples
and add a comment how to configure ISDN interfaces.

The correct way to handle this would be to add support for ispppcontrol
to /etc/rc.network but IMHO we are too close to a release so i will not
touch /etc/rc.network anymore.

Approved by:	murray