Reinitialize the keyboard after the ACPI resume event.
I am not sure if this is absolutely necessary on all systems.  Yet,
there certainly are motherboards and notebook systems which require
this, although there are other systems which just don't.  I hope we
shall know when to do this on which systems, as the development of our
ACPI subsystem progresses...  (I know we didn't need this for the APM
resume.)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@83855 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f