Don't provide an RTA_GATEWAY sockaddr when we write RTM_CHANGE messages to the routing socket.
The local address on a point-to-point interface is not actually a
gateway address - despite it appearing in the second column of
netstat -r's output.  Providing a gateway to an RTM_CHANGE will
currently change the route's interface so that it's using the
specified gateway - not what we want.

Patiently explained to me by:	ru


git-svn-id: svn+ssh://svn.freebsd.org/base/head@87126 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f