Modify our 'patch' utility to recognize the    \ No newline at end of file line that some versions of diff print out if the last line of the two files are different, and one of the two files does not have a newline character on that last line.
This change is still somewhat under discussion in -arch and -standards, but I
want to commit it to -current today so I'd have the chance to MFC it to -stable
before the code freeze for 4.6-release (which would be May 1st).

Note: the related change to 'diff' (so it might *generate* that line) is NOT
expected to be included in 4.6-release.  We can debate that change later.

Obtained from:	NetBSD (1.13 of basesrc/usr.bin/patch/pch.c, by kristerw)
MFC after:	4 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@95514 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f