When freeing a request, zero out the sequence number.
Define the CFG_DAGA_OFF offset as 128 bytes instead of 40- gives us
a more reasonable headroom.

When reading a config page, zero out the entire request area- not just
the length of the request. This is because we cleverly (cheezily) return
configuration data back into the allocated request area, so it's nice
to make sure we start with a clean area to write on.

MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@103827 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f