Fix a refcount race with the vmspace structure.  In order to prevent resource starvation we clean-up as much of the vmspace structure as we can when the last process using it exits.  The rest of the structure is cleaned up when it is reaped.  But since exit1() decrements the ref count it is possible for a double-free to occur if someone else, such as the process swapout code, references and then dereferences the structure. Additionally, the final cleanup of the structure should not occur until the last process referencing it is reaped.
This commit solves the problem by introducing a secondary reference count,
calling 'vm_exitingcnt'.  The normal reference count is decremented on exit
and vm_exitingcnt is incremented.  vm_exitingcnt is decremented when the
process is reaped.  When both vm_exitingcnt and vm_refcnt are 0, the
structure is freed for real.

MFC after:	3 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@107912 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f