Introduce the 	void backtrace(void); function which will print a backtrace if DDB is in the kernel and an explanation if not.
This is useful for recording backtraces in non-fatal circumstances and
does not require pollution with DDB #includes in the files where it
is used.

It would of course be nice to have a non-DDB dependent version too,
but since the meat of a backtrace is MD it is probably not worth it.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@108682 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f