Use floating point instead of unsigned long longs in percentage calculations.  Long longs should never be used, since they break compiling with C90 compilers and don't necessarily work any better than longs for avoiding overflow.
Print percentages with another digit of precision since they can be small
and this is easy to do now that the format is floating point.

Restored some more of the old -m output:
Print the percentage of allocated memory that is in use.  This is the
amount of memory in active mbufs and mbuf clusters relative to the
total amount of memory soft-allocated for mbufs and mbuf clusters.

Print the percentage of allocated memory that is wired (cached).  The
old mbuf allocator never freed memory so printing this value wasn't
useful.  A previous version of netstat for the new allocator printed
the in-use amount as a percentage of the wired amount.

Fixed some nearby style bugs (excessive parenthesization and a redundant
return).

Reviewed by:	alfred


git-svn-id: svn+ssh://svn.freebsd.org/base/head@123958 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f