Modify the implementation of libc semaphores so that they can be overridden by the threads library to provide a userland version of non-pshared semaphores and cancellation points.  Also add a sem_timedwait().
The libc version of semaphores always uses kernel semaphores
regardless of whether pshared is set or not.  When threads are
not present, it is difficult to get sem_wait() or sem_timedwait()
to do the right thing (since pthread_cond_timedwait() and
pthread_cond_wait() are stubs in libc and just return immediately).


git-svn-id: svn+ssh://svn.freebsd.org/base/head@125372 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f