Make withering water tight.
When we orphan/wither a provider, an attached geom+consumer could
end up being withered as a result and it may be in front of us in
the normal object scanning order so we need to do multi-pass.  On
the other hand, there may be withering stuff we can't get rid off
(yet), so we need to keep track of both the existence of withering
stuff and if there is more we can do at this time.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@131820 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f