Free only those mbuf+clusters back to the packet zone that were allocated from there.  All others get broken up and free'd individually to the mbuf and cluster zones.
The packet zone is a secondary zone to the mbuf zone.  There is currently
a limitation in UMA which prevents decreasing the packet zone stock when
the mbuf and cluster zone are drained and all their members are part of
packets.  When this is fixed this change may be reverted.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@152101 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f