Create mediachg functions for the 3c503 and hpp cards.  This is used to properly configure the right interface to use.
Also call the mediachg function when we set flags UP and are already
running.  If this were a pure ifmedia driver, we'd not need to do this
since we'd be ignoring the linkX flags.

This reduces the number of ifdefs to support sub-devices a little as a
nice side effect.  It also reduces the number of hpp interfaces
exposed by 33%.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@154895 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f